start doing migration service

+++ src/Euterpe.Models/Charts/Manifest/ManifestMeta.cs
@@ -15,7 +15,7 @@ public sealed class ManifestMeta
     [Key("bpm_max")] public int? BpmMax { get; set; }
     [Key("scene")] public string Scene { get; set; } = null!;
     [Key("scene_egg")] public string? SceneEgg { get; set; }
-    [Key("background_video_opacity")] public double? BackgroundVideoOpacity { get; set; }
+    [Key("background_video_opacity")] public float? BackgroundVideoOpacity { get; set; }
     [Key("search_keywords")] public string[]? SearchKeywords { get; set; }
     [Key("maps")] public Dictionary<string, ManifestMap> Maps { get; set; } = null!;
     [Key("hide_mode")] public string? HideMode { get; set; }
